+++ BookPlace.Api/Controllers/DownloadLogController.cs
@@ -2,13 +2,14 @@ using BookPlace.Api.DTOs;
 using BookPlace.Api.Extensions;
 using BookPlace.Core.Entities;
 using BookPlace.Core.Interfaces.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BookPlace.Api.Controllers;
 
 public class DownloadLogController(IDownloadLogRepository downloadLogRepository,IBookRepository bookRepository) : BaseController
 {
-    [HttpGet("get-all")]
+    [HttpGet]
     public async Task<ActionResult<IEnumerable<DownloadLogDetailDto>>> GetAll()
     {
         var downloadLogs = await downloadLogRepository.GetAllAsync();
@@ -25,10 +26,12 @@ public class DownloadLogController(IDownloadLogRepository downloadLogRepository,
         return Ok(downloadLogDto);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult> CreateDownloadLog(DownloadLogCreateDto dto)
     {
         var userId = User.GetUserId();
+        if (userId == null) return Unauthorized();
         var book = await bookRepository.GetByIdAsync(dto.BookId);
         if (book == null) return NotFound("Book not found.");
         var downloadLog = new DownloadLog
@@ -42,9 +45,12 @@ public class DownloadLogController(IDownloadLogRepository downloadLogRepository,
         return Ok();
     }
 
-    [HttpDelete]
+    [Authorize]
+    [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteDownloadLog(string id)
     {
+        var userId = User.GetUserId();
+        if (userId == null) return Unauthorized();
         var downloadLog = await downloadLogRepository.GetByIdAsync(id);
         if (downloadLog == null) return NotFound();
         await downloadLogRepository.DeleteAsync(downloadLog);
